Implement basic GA population construction
This implements the construction of a simple population of fixed length GA bit strings, alone with a simple "count ones" fitness function.

I also added rayon's `par_extend` to the construction of the members of the population, and that basically cut the runtime in half according to Criterion, which is very cool! It's only taking ~100 µs for to construct a population of 100 individuals (with the parallelism), so I doubt we'll be able to do a _lot_ to speed this up.

When running in parallel, my laptop does spin up to use all the cores and seems quite busy, especially compared to the serial version where it's clear that only one core is being used.

I should now try to implement essentially the same thing in Clojure and see how they compare.