Set signature for `make_genome` to not be mut
I think that when `make_genome()` "contained" the `ThreadRng` it needed to be `mut`. Now that I pass that in as an argument, it doesn't need to be `mut` anymore, and this reflects that.

I think this is Important and a Good Thing because (I'm hoping) this will make it easier to sort out parameterizing the `Population` type.