Add a helper function `count_ones_vec`
I'm still not entirely sure of the best way to handle the need for an implicit `deref` if I want `count_ones()` to take `&[bool]` and want to pass that in to the parameterized `population::new()` constructor. 

In the meantime, though, adding `count_ones_vec()` as a "pass through" function seems to do the right thing without costing us anything in terms of performance. A bit ugly, but it works.