Fix things so I can use `&[bool]` in argument instead of `Vec<bool>`
It's clear that the recommended approach is to write functions that take slices (like `&[bool]`) as arguments instead of types like `Vec<bool>`, mostly because slices are lighter weight and more flexible.

I was having all sorts of issues, though, where I couldn't pass around functions that took `&[bool]` as an argument because the system wouldn't "realize" it needed to convert a `&Vec<bool>` to `&[bool]` and just to it for me.

@scottmcm on the Rust Discourse (https://users.rust-lang.org/u/scottmcm) shared a solution to this problem that uses the `Borrow` trait: https://users.rust-lang.org/t/problem-passing-functions-as-arguments-and-deref/79491/2?u=nicmcphee

These changes bring those suggestions into my code and actually get everything to work. TBH, I'm not entirely sure I think this is an improvement over just adding the closures that "autoconvert" the types under the covers, but it was certainly interesting.