Add todo note about the problem of multiple fitness computations
Currently if we combine things like mutation and XO, we end up computing the fitness twice, with one of them being thrown away.

We could probably use the Lazy eval tools to deal with that, but I think shifting the bit management (mutation and XO) to the level of bits, and then calculating fitness once when the Individual is actually constructed would make more sense.

I also wonder if there are ways to use the `From` trait here, but I'm not sure we can parameterize that with the fitness function.