Start adding parent selection logic
We got into some pretty weedy parts of Rust here, with some exciting lifetime fun and the use of Higher-Rank Trait Bounds (HRTBs)!

I'm not entirely clear on why we need the boxes, and thus have to move things from the stack to the heap. I think it's because of the `dyn`, but I'm not 100% sure we need that either.

Clippy was pretty excited about the super long type that has here been extracted to `Selector`, which was nice to see. I was also pretty unhappy with how long that was, and it's nice to see that it's happier now.