Add new Selectors-based iterator
This adds an iterator that is built on a vector of Selectors (functions that select an individual from a population).

It turns out that we end up constructing our selectors nearly once for every two individuals created, which seems bad. It wasn't terribly inefficient based on the Criterion benchmarking, but it still annoys me.