Convert from iterator to new `ParentSelector` type
Iterators turned out to be problematic because they encapsulated the random number generator, making them hard to share across threads.

So we created a new `ParentSelector` type with a `get(rng)` method that was similar to the `next()` method  from `Iterator`, but which took the random number generator as an argument. This allowed us to create the parent selector once before going into the threaded code that generated the population in parallel.

I also removed a bunch of commented out code, corrected some comments, and cleaned up the imports.