Move `best_individual` to generalized `Population`
For some reason the `best_individual` logic was in `Population<Bitstring>` when there's nothing there that depends on `Bitstring`s.

This moves it to the general `Population<T>` code where it really belongs.

I also added some `#[must_use]` annotations as recommended by Clippy, and added `is_empty()` and `size()` methods to `Population<T>` that just pass those requests on to the wrapped vector of `Individual`s.