Create new `Generation` type
@esitsu@Twitch suggested creating a `Generation` type that wraps a population and is in charge of things like selection and recombination. That sounded like a good idea, so we spent most of tonight's stream getting that to work.

We currently have a `Generation::par_next()` method to get the next generation from the current one, and we probably ought to make a `Generation::next()` that does the same thing, but uses `into_iter()` instead of `into_par_iter()` for cases where you don't want to take advantage of the parallelism. This would also let us do benchmarking to compare running with and without parallelism, which would be interesting.