Move the `make_child` logic into a named function
I was hoping this would simplify the whole bit with providing types for Clojures, but it turned out that it didn't because we still have to capture the `scorer` function.

This makes me wonder if we just want to explicitly add `scorer` to the `Generation` type, or if there should be a `Run` or `RunParams` type that has all this stuff.