Implement a much faster version of `hiff`
Fiddling with the appending made _some_ difference, but not a ton. I eventually decided that the real problem was allocating and copying all that memory/data in the recursive calls.

So I created a `scores` array and passed it around and everyone just wrote into the appropriate slots in that array, so there was no additional allocation or copying.

This made a _big_ difference in the calculation of `hiff`, speeding things up by a factor of more than 10. That translated to an improvement of around a factor of three for full runs of the GA.