Simplify definition of HIFF
This rewrites `hiff` so that it uses `push` directly on the scores vector instead of keeping track of the current index and all that weirdness. This may be a hair slower, but only a hair, and it's definitely easier to read and less error prone, so I think I prefer it.