Add two new versions of lexicase selection
`one_pass_lexicase` tries to translate the approach taken in Ellen over here to Rust. There are actually two versions of that; one uses an explicit for-loop and the other uses `fold()`, but it didn't make a bit difference to the performance.

`reuse_vector_lexicase` is similar to the previous one, but uses some nifty tricks (thanks to esitsu@Twitch for the `swap` idea!) to minimize the allocation and deallocation.