Add test results type to everything
This started by adding the new generic type `R` for the test results to `Individual`, and that then spread out through pretty much all of the other code, which was kinda gross. Lots of grinding through errors, but in the end it wasn't _too_ bad. Not what I would call fun, though.