Move test results to new module; implement `from` & `sum` traits
This moves all the test results types (`Error`, `Score`, `TestResults`, etc.) into their own `test_results.rs` module, thus tidying up `individual.rs` a fair bit.

This also includes what is arguably the big win of the day, which is implementing `From` and `Sum` traits on `Score` and `Error`, and a nifty implementation of the `Sum` trait for `TestResults`.

In retrospect, I do think that `Sum` wasn't the perfect trait to implement for `TestResults` since a `TestResults` value isn't really the "sum" of a bunch of individual errors. `.collect()` from the `FromIterator` trait is probably the better choice, but that wouldn't be hard to do (& `Sum` is working well for now).