Implement `Weighted` composite selector
This replaces the old `weighted_selector` field type with a new composite implementation of the `Selector` trait. This lets us either give things like `Generation` either a "simple" selector (like `Lexicase`) or a composite selector (e.g., `Weighted`), and nothing has to change in `Generation`.

We also used the builder pattern for `Weighted`. Since the collection of selectors should never be empty, I implemented `new()` to take a single selector-weight pair, and then `with_selector` allows us to add additional selector-weight pairs.

This was one of the high points of today's stream and worked beautifully!