Make `VecPop::individuals` private
In preparation for making a `Population`, we made the `individuals` field in `VecPop` private. This cleaned up several direct uses of the `individuals` when calls to `is_empty()` or `size()` would have been more appropriate.

We had to add an `iter()` method and a pair of `FromIterator` implementations to be able to convert back and forth between collections of `Individuals`.