Add `ChildMakerI<I>` trait
This adds a new `ChildMakerI<I>` trait that is only generic in a single type instead of the `<G, R>` pair for the old `ChildMaker` trait. We also moved this to its own module file.

To remove a circular reference between `Generation` and `ChildMaker`, the `make_child_i` method in the new interface takes the population and the selector (from the `Generation`) instead of taking the generation directly.

We had to make several small changes to the `TwoPointXoMutationChildMaker` to match these changes.