Make `Generation` generic in three types
This makes `Generation` generic in types for the population, the selector, and the child maker.

This also required that `ChildMaker` be generic in a type for the selector.

In the process, `Generation` no longer holds _references_ to the selector and child_maker, which is why they need to be clonable.