Implement Selectors & ChildMakers as refs
This adds implementations for both `&dyn Selector` as `Selector`, and `&dyn ChildMaker` as `ChildMaker`. This allows us to pass in references to selectors and child makers to `Generation`, and they will work fine. The nifty thing is that when we then clone the selector in child maker when going from one generation to the next, we're just cloning the references, which is nice and cheap. :)

We needed the `+ Send + Sync` versions for `ChildMaker` because that needs to be shared between threads when making a new generation. (Not sure why that isn't true for `Selector`, TBH.) esitsu@Twitch suggested that we provide three versions for each: +Send, +Send+Sync, and neither. This provides a lot of flexibility, and follows the model set in the `Any` trait (see the `downcast` signatures in https://doc.rust-lang.org/std/any/trait.Any.html#method.downcast).

esitsu@Twitch suggested having