Use `Self` in `TwoPointXoMutate` impl
I'd used `TwoPointXoMutate` in two places where I could have used `Self`, which became clear when I copy/pasted that code into the new `Pipeline` implementation of `ChildMaker`. This fixes that.

This required adding a lifetime. I'm not entirely sure why it wasn't necessary before, but presumably when we had the explicit types the compiler could infer the lifetime in a way that it now can't.