Add `operator.rs`
I've decided (for now) to go with esitsu@Twitch's suggested recombinator design.

To make sure I understood what was happening, I created `operator.rs` and started by implementing the `Operator` trait and the `Then` implementation of that trait to see how the types would play out.

We then added `Mutate` and `Mutator` traits, `Select` and `Selector` traits, and a `Compose` trait with `then_mutate()` and `and`select()` methods.

It's a start, but there's still a _lot_ of work to do.