Implement several operators as `Operator`
This implements several of the existing "operators", i.e., things that implemented `Recombinator`, as `Operator` (and by extension `Composable`):

- `MutateWithRate`
- `MutateWithOneOverLength`
- `TwoPointXo`

We also update `TwoPointXoMutate` child maker to use these new operators instead of the the recombinators.

I actually "kept" all the `Recombinator` implementations by having them call the new `Operator` implementations. Once we've moved everything to be `Operator`, though, then we want to get rid of `Recombinator` altogether.

I've marked `Recombinator` as `deprecated`, but that generates a _lot_ of noise, so I may undo that until we've farther along in the process of replacing it.