Move `Sized` into `then` in `Composable`
Instead of making all implementations of `Composable` implement `Sized`, this moved the `Sized` requirement into the function so it only is checked/required when we use the `then` function.

This is related to the whole object safety thing that I don't understand very well and need to do some homework on.