Remove all the references to `Selector` in `Weighted`
This changes the vector of selectors to be impls of the `Operator` trait instead of the `Selector` trait, and removes all references to `Selector` except for the one `impl Selector` that we'll ultimately remove.

When we do remove that, we should be able to simplify some of the existing code (e..g, removing the `'pop` lifetime in `SelectorOperator`, and make more general.