Change `ChildMaker` to use `Operator` instead of `Selector`
This changes the `ChildMaker` trait to use `Operator` (with the appropriate "selection" aspects) instead of `Selector` as part of eventually removing the `Selector` trait.

In `lib` we passed the selector and child maker in as owned values instead of as references, which simplified a number of things.

In the process we removed `Generation::get_parent()` because it turned out that we never called it anywhere.