Added esitsu's wrapper type for `Selector`
I really wanted to have a name for selection operators instead of just having the huge generics. We couldn't do that with a `Selector` trait, which brought us back (after many weeks) to esitsu@Twitch's (now old) suggestion to have wrapper types to allow us to move from `Selector` to `Operator`.

So here we have a new `Selector` trait and specifies a `select` method. We then make a wrapper `struct Select` that implements the `Operator` trait which is specialized for selection. This allows us to encapsulate all the weird generics in `Select`, and also `impl Composable for Select` so we don't have to `impl Composable` for all the specific selectors.