Convert `Weighted` to `Selector`
This was a little more work because we had to figure out what to store in the vector of selectors. We went with `Vec<(Box<dyn Selector<P>>, usize)>` at first, although we ended up having to add `+ Send + Sync` to the boxed selectors so that they can be passed around among threads.

I think we could have use references instead of `Box` here, but references would have required lifetimes and complicated things in other ways. I'm not sure what the performance implications are (if any), but I must say that using `Box` instead of lifetimes is easier to read.

Once we'd made those decisions, the rest of the changes were pretty straightforward.