Add `Map` as a composable operator
This adds a new `Map` operator that takes a function (as an `Operator`) and applies it to some collection of values, returning a new collection of values.

This also adds a `.map()` method to the `Composable` trait so we can use this to map operators across collections of values in pipelines.