Add variants of the recombinators that take pairs instead of vectors
Both my recombinators `TwoPointXo` and `UniformXo` took vectors, but when we built the pipeline, `And` returned a pair. So I've added variants of the recombinators that take pairs and just call the associated method that takes a pair instead. 

This is perhaps a little icky. It potentially introduces a new function call, although I don't know if the compiler can inline that or not.

I should figure out what I want my pipeline components to actually take and return, and whether I want vectors or tuples for things like this. Once I have that figured out, I should then change these so the "common" one doesn't involve the extra function call.