Extend the pipeline in `TwoPointXoMutate`
This extends the pipeline in `TwoPointXoMutate` so that it now goes from the previous population all the way to the mutated genome in one pipeline!

This also makes the old `make_child_genome` function obsolete, so I removed it. That pointed out that the "smoke test" was really just testing a small part of the pipeline (constructing the child genome from parents), so I updated the test to reflect that.