Resolve the `TwoPointXoMutate` compilation issues
One simple was that in `lib.rs` we were passing a reference to a `scorer` instead of taking ownership when we constructed the `TwoPointXoMutate`.

The key piece of the fix in `TwoPointXoMutate` was changing the `Sc` trait bound to be `Fn(&[bool]) -> Vec<i64>`. It _had_ to match the type of the scoring functions (e.g., `count_ones`) so the input had to be a slice and the output had to be `i64`s.

The other thing we _had_ to do was specify the type of the input to the `make_test_results` closure to be `&Vec<bool>`. Without that, the definition of the closure led the compiler to infer the input has having the type `&[bool]`, which then didn't match with the `apply` call below. If we specify the type as `&Vec<bool>` then when the `self.scorer` is called, then the `&Vec<bool>` is converted to a `&[bool]` and the rest works.

The compiler was not very helpful here, and I have no idea how I would have resolved without the greatly appreciated assistance of esitsu@Twitch who put a lot of effort into this.