Add comment to `Recombinator` trait
I've been bad and not commenting my traits as I've gone along, and this has meant that I've often been confused and had to put more effort into re-learning what code (that I wrote!) did than I should have. I'll have to try to improve on that moving forward, and this is an example of that.