Add the new `Crossover` trait
esitsu@Twitch suggested that I add a `Crossover` trait as a sub-trait of `LinearGenome` that essentially specifies the key operations needed for performing crossover:

- Swapping a single gene at a given location
- Swapping segments of genes specified by a `Range`.

This made it easier to implement the actually crossover operators (e.g., `TwoPointXo`) for `LinearGenome`s.

I'm not sure I would have thought of creating this trait, but it does do a nice job of factoring out the "gene level" crossover operations that will be shared by most crossover operations on fixed length structures.