Change `gene.rs` to `genome.rs`
The `gene.rs` module was really far more about `Genome`s than `Gene`s (and had the `Genome` trait`), so I renamed the module.

We added a new `Generator` trait (suggested by #esitsu) that takes a context and returns a (probably randomly generated) value of type `T`. The context allows us to provide other information that might be necessary, like a distribution that values should come from or the length of a bit string or size of a population.

We also added an implementation of `Generator` that generates random `bool`s with a given probability.

The somewhat odd thing is that we're implementing the `Generator` trait on `ThreadRng`, with the return type and the context specifying what kinds of things we're generating and how. So we essentially get a bunch of different `rng.generate()` methods, where the context argument and the return type giving us something like polymorphism and allowing the compiler to know which specific `generate()` to use.