Add `impl Generator for Population`
This was pretty straightforward. I tried parallelizing it in `Population` using Rayon, but that got a little awkward because we've attached `Generate` to `ThreadRng`. `ThreadRng` can't implement `Send` since the whole point is to _not_ be sharing it across threads, so you can't pass `self` to each thread. I did get it to compile (at least) with calling `thread_rng()` on each thread instead of using `self`, but it did complicate things so I removed it in the end. Maybe what we want is to have both the parallel and the serial versions available, with the choice specified by a boolean (or `enum`) in the `population::GeneratorContext`?