Change the test functions to return `TestResults`
This changes `count_ones` and `hiff` to return `TestResults<i64>` instead of `Vec<i64>`, which simplifies a number of things in `main`. I'm not 100% I'm thrilled about that decision, TBH, but I guess it doesn't make sense to return a `Vec<>` if you're not going to (ultimately) turn it into a `TestResults<>` value.

This was simplified by adding a `From<>` trait that converts from `Vec<V>` to `TestResults<V>`, so we can just use things like `.into()` to convert from a `Vec<>` to `TestResults<>`.

I also deprecated `BitstringVecType` since the goal is to get rid of that.