Implement `Generator` for `EcIndividual`
This implements the `Generator` trait for `EcIndividual`, using the new `Scorer` trait.

I'm guessing that the `context.scorer.score(&genome)` call is an interesting example of a zero-cost abstraction. While it looks like there's a lot happening there, my guess is that the compiler turns that into a simple function call when the scorer is a simple function like `count_ones`.