Write `Generator` tests for `Population`
I had `Generate` tests for `Population`, and I wanted to also have some tests for the implementation of `Generator`, especially since the plan is to get rid of all the `Generate` stuff.

One consequence of moving `population.rs` into `ec-core` is that we no longer have access to the new `struct BitString`. So for these tests I created a simple `RandValue` type that used a `Range` as its context.