Change `count_ones` and `hiff` to return `Score`
Instead of having `count_ones` and `hiff` return `TestResults<i64>`, this changes them to return `TestResults<test_results::Score>`, which was a suggestion from esitsu@Twitch. At first I was a little resistant because I didn't want to "complicate" these functions, but after I'd had a little time to think about it I decided that this was a pretty good idea. It then pushes the decision about whether a fitness function is returning an `Error` or a `Score` to that fitness function, where it arguably belongs. And then there's clarity about that all up and down the chain.