Simplify `Generation`
Since the `Selector` is now in the `child_maker`, we don't need that as a generic anymore.  And now that `child_maker` is an `Operator`, we use `apply` instead of `make_child` in the `next()` functions.

This also changes the `next()` functions to be "in place", i.e., they replace the original population _in this generation_ with a new population and don't actually return anything. This potentially allows us to avoid a lot of allocation/deallocation.