Get `main()` to work using the new structure
This gets `main()` working with the new `Operator`/`Generation` system. 

`make_new_individual` is now a "pure" chain of `Operators` using several different composition operators to put it all together.

Man, there are a lot of dependencies in the `use` clauses here. I guess that's to be expected, but it definitely is an awful lot. I wonder if that's common in Rust projects, where the dependency tree is shallow and wide, especially since you don't have the mirage of sub-typing giving you deep dependency trees.