Replace `.sum()` with `.into()` for `TestResults`
I was using `.sum()` to convert a vector of errors/scores into a `TestResults` value, and that always seemed confusing to me. This changes that so that `TestResults<R>` implements `From<I>` where `I` is an iterator over values (e.g., scores or errors), as well as implementing `FromIterator` (which allows us to use `.collect()` if we already have an iterator.

I think makes a lot more sense, and cleans things up quite a bit.

I also added implementations of `Sum` for _references` to `Score`/`Error` so that we can compute sums without taking ownership.