Clean up uses of conversions to `TestResults`
After all the changes I made to the tools for converting to `TestResults`, I hadn't really used them fully in other places.

This goes through all the places where we do that conversion, and try to rationalize what's happening there.

The coolest change is in `ec-push-cli/main`, where I was able to remove the bit that converted integral values to `Error`s since the conversion tools will do that bit for me.

I also changed the signature of `run_to_completion` so that it returns a reference to the `State` so we can chain `run_to_completion()` with additional calls (like we now do in `main.rs`).