Simplified the derived traits for `Score` and `Error`
It turns out that neither of them needed `Copy` or  `Clone` anymore, so I removed those from the `derive` clause.

We may want to put those back in later for generality, but it was rather nice to find that I didn't actually need either of them. I think that the changes that I made to the construction of `TestResults` values changed things so I don't need those anymore.