Rewrite `Inputs` system and `PushState::Builder`
This came from a combination of ideas from MizardX@Twitch and esitsu@Twitch.

MizardX had suggested on Discord that we should try to replace the string equality tests in `PushState::push_input()` with integer equality tests. I spent a bunch of time wrestling around the issues there, and eventually esitsu suggested that maybe we wanted to look at the Typestate pattern. I did a little reading quick (https://www.greyblake.com/blog/builder-with-typestate-in-rust/) and decided that looked like a cool approach.

We then got that sorted out, and I'm pretty happy with how that now works.