Add new `LinearContext` for generating random linear genomes
This new context allows us to generate random linear sequences (e.g., `Vec<T>`) given a context `C` that can be used to generate the elements of type `T`.

We then use this to replace the code for generating `Bitstring` so they use `LinearContext<BitContext>` to generate the necessary Bitstrings.

That required making changes to several tests and the `ec-ga-cli::main()`.