Implement the `Plushy` genome
This replaces the "raw" `Vec<PushInstruction>` with a new `Plushy` genome type and associated random `Generator` implementation.

It was crucial to change the `scorer` closure in `main()` to take a `&Plushy` instead of a `&Vec<PushInstruction>`. It took me a while to realize that I needed to change that, and until I did we were still generating populations of `Vec<PushInstruction>` genomes instead of populations of `Plushy` genomes.