Make `CollectionContext` clonable
I need this so I didn't have to make two copies of the `CollectionContext<PushInstruction>` in `ec-push-cli::main()`. Perhaps it would make more sense to make the contexts references instead of owned types so we don't need to clone them? We only clone it once, though, in the run setup, so it's hardly a major performance concern.