Write a simple smoke test for `Umad`
This provides a simple smoke test for `Umad` that makes sure that we can run `Umad` on a `Vector<char>` linear genome. We check that it actually inserts a few new characters, deletes a few characters, and that the remaining characters are a subsequence of the original string.

This is a stochastic test, so it will fail occasionally, but not too often.