Rename `LinearGenome` to `Linear`
Clippy was fussing about this, and at first I wasn't super keen, but I tried it and found that I quite liked hot things like `impl Linear for Bitstring` read, so I kept it.

If I ever need it in a context where it's not clear what's happening, I can always use `genome::Linear` as a more qualified type name to improve readability.