Fix the use of `then_map` in `main()` and switch to `for` loop
The compilation problem in `main()` was because I had copied `then_map(GenomeExtractor)` from the `Bitstring` example, which was wrong here because we had a single parent instead of a pair of parents, so no mapping was required (or allowed). Thanks to esitsu@Twitch for highlighting the fact that `Map` was perhaps the culprit.

We also changed the loop from using `try_for_each` to an old-school `for` loop so we could use `break` to exit early when we found a solution.