Refactor `Generator` to attach to contexts instead of rngs
This is a pretty substantial refactoring proposed by esitsu@Twitch on Discord. The key change is going from having `ThreadRng` implement `Generator` for different contexts, to having the contexts implement `Generator`, passing in the random number generator to their argument. This simplifies a number of things, including the fact that you can't share `ThreadRng`s across threads, but you can share `Context`s across threads. This took something over an hour, but I think it was a valuable change and I'm glad we did it.

We could extend this idea to have `Generator` be generic in some type `R` that is passed to the `generate()` method. In most circumstances that would just an a `ThreadRng`, but we could pass other things (like other random number generators) if we wanted to.