Set up calling Rust `lexicase` from Python
This gets that to work! esitsu@Twitch shared this article (https://ohadravid.github.io/posts/2023-03-rusty-python/), which was an example of exactly the kind of thing we were trying to do, which made a _huge_ difference in getting this to work.

At this point we only have `select_one()` implemented, so we still need to implement `select()`. I'm guessing that when we do, if we need to select many parents (esp. on the order of 100s or a thousand), then the Rust implementation will win pretty handily.

Because we're using the version of `Lexicase` from the (quite general) space in `rust-ga`, we have to do a lot of conversion (which isn't cheap) to get things in the right form to call `Lexicase::select()`. If we copied the logic of `Lexicase` out into this package and converted it to match the PyshGP types then we'd avoid all that conversion and presumably speed things up a fair bit.