Ignore the Python `.env` file
Setting up the Python tools created a huge `.env` directory. I'm not sure what parts of that people recommend committing in `git` and what they don't, but given that I'm not sure we'll actually keep any of this Python stuff in the system, I'm going to be conservative and ignore it for now.