Add all the `int` and `bool` instructions that are in Clojush
This fleshes out the `int` and `bool` instruction sets. I used the Clojush instruction sets as the model here, and added everything that made sense from those. There are a few that I only mentioned in comments because they needed a stack we didn't have (`BoolFromFloat`), or I wasn't sure we actually wanted them (`BooleanInvertFirstThenAnd`).

If we're going to use Rust, we have potential overflow issues on several of the int operations. I'm not sure how we want to handle that. At the moment I just skip the action and push the results back on, but I doubt that's actually a great solution.

It probably makes more sense to have instructions return an `Option` or a `Result`, and then have the state system only update the stacks if we get a `Some()` or `Ok()` result?

There's also a ton of duplicate code, here, and I'm sure we can do better. Not sure if that's a matter of extracting functions, or if we should be using macros for that.