Add `proptest` tests for `int` and `bool` instructions
This adds `proptest` tests that don't check the _correctness_ of the integer and boolean instructions, but it does check that they don't crash. These tests highlighted issues with many of the integer instructions (add, multiply, sqr, etc.).

I have a test for the logic of one instruction, `Add`, where I check that the result is the correct addition if that succeeds, or is still the original top value if adding them would overflow. I arguably want similar tests for the other instructions, but that is going to be tedious and involve a ton of duplicate code unless I do something clever.

I wonder if I want to further organize the int instructions into unary and binary instructions, or maybe instructions that might overflow and those that might not?