Rename `PushStack` to `HasStack`
It was clear that `PushStack` was a terrible name for that type (we were very confused about it in the stream), but I wasn't sure what to call it. Now that the design has started to settle, I think that what this does is say that we have/can access a stack of a given type. I'm not sure if `HasStack` is a great name, but it does seem to convey the intent so we'll go with it for now.