Finish implementation of `Stack<T>`
This finishes the implementation of the methods in `Stack<T>` that I'd left as `todo!()` at the end of the stream.  I also added `size()` and `top()` methods while I was there.

As part of this I cleaned up a few Clippy warnings along the way, and ended restructuring `run_to_completion()` so that it takes ownership of the state and returns the updated state instead of modifying a mutable reference.