Add tools to initialize int and bool stack values
This adds an `extend()` method to `Stack` and `with_bool_values()` and `with_int_values()` to the `PushState` `Builder` so we can initialize our states with given sequences of values.

I then used those in restoring the `proptest`s for `BoolInstruction`. While I was there I also added a `BoolImplies` instruction and added a test for it. It's probably good to have that instruction, and I wanted something that wasn't symmetric so I could make sure I as adding values in the right order.

Since I always find the stack order confusing, I added doc comments for `extend` and `with_*_values` that are explicit about the order that things will go on the stack, including runnable examples.