Change handling of input variables
I found the whole business with inputs was confusing and awkward, so I managed to change things so that we just use a simple `HashMap` from variable names (which I'm modeling as `Arc<str>`) to `PushInstruction`s. I think that this is much simpler and easier to make sense of. It is a bit slower than the previous system, especially if I use a `HashMap` since the initialization of that `HashMap` is more expensive than initialization of `Vec`. If there are a lot of variables, then the linear lookup in `Vec` will hurt, but if there are only a few then `Vec` seems to be better. My guess is that this doesn't really matter since initializing the state is a pretty small price compared to actually running the programs.