Switch from `Vec` to `HashMap` for input instructions
`Vec` had been a bit faster when the programs were fairly short, but when I bumped up the length of the programs considerably to better model the execution times, then using `Vec` and `HashMap` had about the same execution. So I switched over to `HashMap` since that will scale better in cases where there are lots of variables.