Have instructions own state and return `Result`s
This is a start on having the execution of `Instruction`s take ownership of the state and return a `Result` type that allows us to capture information about events like numeric overflow and (eventually) exceeding step limits.

Not everything is done (the code doesn't compile yet), but it's a useful start.