Add `StackError` type
One nice thing about using `thiserror` is that it encouraged me to think in terms of hierarchies of error types. This led to thinking of `StackError` as a separate, lower level, kind of error, with `PushInstructionError::StackError` being derivable `#[from]` a `StackError`.

I also added a `max_stack_size` value to `Stack` as a way of checking if we've reached the maximum stack size for a stack.