Radically improve `BoolInstruction::perform()`
Converting `BoolInstruction::perform()` to return the new `InstructionResult` type enabled a huge improvement in the nasty logic there. Many thanks to esitsu@Twitch for asking the questions that ultimately led to these improvements.

The key thing is to use `.map()` to process the values returned by `.pop()` or `.pop2()`, and then `.map_err()` to convert `StackError`s to the appropriate `PushInstructionError`s.