Add `.stack()` to `HasStack`; move `PushInteger` type to `push_state.rs`
I added a `.stack()` to the `HasStack` trait that returns a non-mutable version of a typed stack. I then went through all the calls to `.stack_mut()` and replaced those that didn't require mutability (usually things that were reading results from a stack with `.top()`) with `.stack()`.

I also moved the `PushInteger` type from ? to `push_state.rs` since I think it makes more sense to associate that with the state generally rather than with `IntInstruction`.