Add comment about Stacks being transactional
In the current implementation, stack operations need to be "transactional", i.e., all the operations have to succeed or none of them are performed. E.g., `pop2()` checks the size of the stack, and only pops anything if it can get them _all_, so it either completely succeeds, or it doesn't change the stack and at all and returns an error.

In truth, however, this is only partly the case. Instructions like `IntInstruction::FromBool`, for example, pop from the boolean stack and then (later) push onto the integer stack. It's possible that the pop succeeds, but that the push fails, and we end up an inconsistent state where the boolean has been removed but the int hasn't been pushed.