Add new `MapInstructionError` trait
This let's us convert the error type _inside_ an `InstructionResult` to a new error type. This was needed because `instruction::Error` wraps an error value, and it's the _wrapped_ error value that we want to map into a new type, leaving the rest of the `instruction::Error` alone. We tried implementing a `From` somewhere to automate that process, but it failed because we needed to implement `From` on `Result`, and we can't do that because we don't "own" either of those types. Now that I write these words, though, it's clear to me that I don't _really_ understand what was going on here, and I'm wondering if there was a better way. One question is whether I want to make `InstructionResult` an actual `struct` instead of just a type alias. then I could have implemented `From` on it, and that might be cleaner?