Add `TypeEq` trait to allow typed turbofish on `HashStack`
Justus_Fluegel@Twitch shared (last week) his `TypeEq` trait, which allowed us to have functions like `HasStack::stack()` be parameterized by the type of the stack that we want. This allows us to say things like `state.stack::<bool>()` to get a reference to the boolean stack without having to specify the type of the stack somewhere else like in the type of the variable we assign this to.

I think this is pretty nifty, but definitely uses some pretty serious Rust voodoo to work. :-)