Start updating the integer instructions
We have a lot of work still to do here, but it's a start on updating these to use `MapInstructionError`, `StackPush`, and the updated `HasStack`.

Unfortunately (as shown in `Inc`), there's still a lot of boilerplate wrapping and obscuring the important logic. I think maybe there's another trait somewhere to collect all that up.